Add support for termination signal dispatch.
Add support for terminating all processes of the timeout
group when a terminating signal is sent in addition to when
a timeout occurs.
By default the list of terminating signals ar INT, TERM, QUIT and
can ba changed with the -c option.
Added -l PID and -f FORMAT options for respectively listing the
processes in the given timeout group and formating the list.